add additional accounts inside params-estimator
