+++ genesis-tools/genesis-populate/src/main.rs
@@ -6,7 +6,7 @@ use clap::{App, Arg};
 use near_store::create_store;
 use nearcore::{get_default_home, get_store_path, load_config};
 
-use genesis_populate::GenesisBuilder;
+use genesis_populate::add_additional_accounts;
 
 fn main() {
     let default_home = get_default_home();
@@ -26,15 +26,5 @@ fn main() {
         .value_of("additional-accounts-num")
         .map(|x| x.parse::<u64>().expect("Failed to parse number of additional accounts."))
         .unwrap();
-    let near_config = load_config(home_dir);
-
-    let store = create_store(&get_store_path(home_dir));
-    GenesisBuilder::from_config_and_store(home_dir, Arc::new(near_config.genesis), store)
-        .add_additional_accounts(additional_accounts_num)
-        .add_additional_accounts_contract(near_test_contracts::tiny_contract().to_vec())
-        .print_progress()
-        .build()
-        .unwrap()
-        .dump_state()
-        .unwrap();
+    add_additional_accounts(home_dir, additional_accounts_num);
 }
